+++ tests/testshared.h
@@ -159,7 +159,7 @@ typedef struct test_timing_s {
 #define TIMING_ASSERT0(TIMING,DO_FAIL_LATER,TMIN_MS,TMAX_MS) do {       \
         if (!TIMING_STOPPED(TIMING))                                    \
                 TIMING_STOP(&timing);                                   \
-        int _dur_ms = TIMING_DURATION(TIMING) / 1000;                   \
+        int _dur_ms = (int)TIMING_DURATION(TIMING) / 1000;              \
         if (TMIN_MS <= _dur_ms && _dur_ms <= TMAX_MS)                   \
                 break;                                                  \
         if (test_on_ci)                                                 \
